+++ app/models/admin.rb
@@ -8,6 +8,7 @@
 #  site_id       :integer       
 #  template_name :string(100)   
 #  flags         :text          
+#  super_admin   :boolean       
 #
 
 class Admin < ActiveRecord::Base
@@ -86,5 +87,6 @@ class Admin < ActiveRecord::Base
     manage_sync
     run_reports
     manage_reports
+    manage_contributions
   )
 end