+++ app/models/group.rb
@@ -29,6 +29,7 @@
 #  gcal_private_link         :string(255)   
 #  approval_required_to_join :boolean       default(TRUE)
 #  pictures                  :boolean       default(TRUE)
+#  cm_api_list_id            :string(50)    
 #
 
 class Group < ActiveRecord::Base
@@ -59,6 +60,7 @@ class Group < ActiveRecord::Base
   validates_format_of :address, :with => /^[a-zA-Z0-9]+$/, :allow_nil => true
   validates_uniqueness_of :address, :allow_nil => true
   validates_length_of :address, :in => 2..30, :allow_nil => true
+  validates_uniqueness_of :cm_api_list_id, :allow_nil => true, :allow_blank => true
   
   serialize :cached_parents
   
@@ -115,11 +117,6 @@ class Group < ActiveRecord::Base
     write_attribute(:address, a == '' ? nil : a)
   end
   
-  def leader_with_guessing
-    leader_without_guessing || admins.first
-  end
-  alias_method_chain :leader, :guessing
-  
   def get_options_for(person)
     memberships.find_by_person_id(person.id)
   end
@@ -142,6 +139,9 @@ class Group < ActiveRecord::Base
       update_membership_associations(Person.find(:all, :conditions => conditions))
     elsif Membership.column_names.include?('auto')
       memberships.find_all_by_auto(true).each { |m| m.destroy }
+    end
+    if cm_api_list_id.to_s.any? and Setting.get(:services, :campaign_monitor_api_key).to_s.any?
+      sync_with_campaign_monitor
     end
   end
   
@@ -227,10 +227,12 @@ class Group < ActiveRecord::Base
     Group.find_all_by_parents_of(id).each { |g| g.update_attribute(:parents_of, nil) }
   end
   
-  def sync_with_campaign_monitor(api_key, list_id)
-    cm = CampaignMonitor.new(api_key)
+  def sync_with_campaign_monitor
+    return unless (api_key = Setting.get(:services, :campaign_monitor_api_key)).to_s.any?
+    return unless cm_api_list_id.to_s.any?
+    cm = CampaignMonitorParty.new(api_key)
     in_group = self.people.all(:conditions => ['memberships.get_email = ?', true]).select { |p| p.email.to_s.any? }
-    unsubscribed = cm.Subscribers.GetUnsubscribed('ListID' => list_id, 'Date' => '2000-01-01 00:00:00')['anyType']['Subscriber'].to_a.map { |s| [s['Name'], s['EmailAddress']] }
+    unsubscribed = cm.Subscribers.GetUnsubscribed('ListID' => cm_api_list_id, 'Date' => '2000-01-01 00:00:00')['anyType']['Subscriber'].to_a.map { |s| [s['Name'], s['EmailAddress']] }
     # ensure we don't resubscribe someone who has already unsubscribed
     # (and also set their get_email attribute to false in the group)
     upload_to_cm = []
@@ -243,16 +245,16 @@ class Group < ActiveRecord::Base
       end
     end
     # unsubscribe addresses in the subscriber list but not found in the group
-    subscribed = cm.Subscribers.GetActive('ListID' => list_id, 'Date' => '2000-01-01 00:00:00')['anyType']['Subscriber'].to_a.map { |s| [s['Name'], s['EmailAddress']] }
+    subscribed = cm.Subscribers.GetActive('ListID' => cm_api_list_id, 'Date' => '2000-01-01 00:00:00')['anyType']['Subscriber'].to_a.map { |s| [s['Name'], s['EmailAddress']] }
     subscribed.each do |name, email|
       if not upload_to_cm.any? { |n, e| e == email }
-        cm.Subscriber.Unsubscribe('ListID' => list_id, 'Email' => email)
+        cm.Subscriber.Unsubscribe('ListID' => cm_api_list_id, 'Email' => email)
       end
     end
     # subscribe addresses in the group but not in the subscriber list
     upload_to_cm.each do |name, email|
       if not subscribed.any? { |n, e| e == email }
-        cm.Subscriber.Add('ListID' => list_id, 'Email' => email, 'Name' => name)
+        cm.Subscriber.Add('ListID' => cm_api_list_id, 'Email' => email, 'Name' => name)
       end
     end
   end