+++ app/models/news_item.rb
@@ -87,14 +87,18 @@ class NewsItem < ActiveRecord::Base
     end  
     
     def get_feed_items
-      if url = Setting.get(:url, :news_feed)
+      urls = []
+      urls << Setting.get(:url, :news_feed) if Setting.get(:url, :news_feed).to_s.any?
+      urls << "#{Setting.get(:services, :sermondrop_url).sub(/\/$/, '')}/sermons.rss" if Setting.get(:services, :sermondrop_url).to_s.any?
+      urls.map do |url|
+        next unless url.to_s.any?
         begin
           feed = Feedzirra::Feed.fetch_and_parse(url)
           feed.entries
         rescue
           nil
         end
-      end
+      end.flatten.compact
     end
   end
 end