+++ config/environment.rb
@@ -24,13 +24,13 @@ Rails::Initializer.run do |config|
   config.gem 'liquid',                 :version => '~> 2.0.0'
   config.gem 'highline',               :version => '~> 1.5.0'
   config.gem 'mini_magick',            :version => '>= 1.2.5'
-  config.gem 'will_paginate',          :version => '~> 2.3.0'
   config.gem 'campaign_monitor_party', :version => '>= 0.2.1'
   config.gem 'whenever',               :version => '~> 0.4.0'
   config.gem 'nokogiri',               :version => '~> 1.4.0'
   config.gem 'builder',                :version => '>= 2.1.2'
   config.gem 'feedzirra',              :version => '>= 0.0.20'
   config.gem 'mongo',                  :version => '>= 0.18'
+  config.gem 'fastercsv',              :version => '~> 1.5.0'
 end
 
 (Setting.update_all if Setting.table_exists?) rescue nil