+++ lib/inherited_attribute.rb
@@ -19,7 +19,10 @@ module ActiveRecord
         class_eval \
           "
           def share_#{attribute}_with(person)
-            visible? and (
+            read_attribute(:visible) and
+            (!respond_to?(:family_id) or
+              (family and family.visible?)
+            ) and (
               share_#{attribute}? or
               self == person or
               (self.respond_to?(:family_id) and self.family_id == person.family_id) or